+++ src/components/daybook/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const DaybookPage = () => {
-  return <div>DaybookPage</div>;
-};
-
-export default DaybookPage;