+++ src/stores/store.ts
@@ -66,8 +66,7 @@ const expiringStorage = {
 
 export const useStore = create(
   persist<StoreState, [], [], PersistedState>(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    (set, get) => ({
+    (set) => ({
       admin: null,
       coldStorage: null,
       token: null,